refact(src): split files over your current header and put a Makefile
