+++ src/headers/utils.h
@@ -1,0 +1,8 @@
+#ifndef UTILS_H
+#define UTILS_H
+#include <stdio.h>
+
+void swap(unsigned char* firstElement, unsigned char* secondElement);
+void printRotor(unsigned char* arr);
+
+#endif
